% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{angle}
\alias{angle}
\title{Calculates angle of two vectors.}
\usage{
angle(x, y)
}
\arguments{
\item{x}{vector 1.}

\item{y}{vector 2.}
}
\value{
angle in radian.
}
\description{
Calculates angle of two vectors.
}
\examples{
# a example of orthogonal vectors
angle(c(0, 1), c(1, 0))
# = pi / 2

}
