% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{arrows.fill}
\alias{arrows.fill}
\title{Draws arrows between pairs of points with filling arrow head.}
\usage{
arrows.fill(x0, y0, x1, y1, length = 0.1, length.fixed = F, angle = pi/6,
  ...)
}
\arguments{
\item{x0}{coordinates of points from which to draw.}

\item{y0}{coordinates of points from which to draw.}

\item{x1}{coordinates of points to which to draw.}

\item{y1}{coordinates of points to which to draw.}

\item{length}{length of the edges of the arrow head (ratio towards length of arrow).}

\item{length.fixed}{if \code{T}, argment \code{length} indicates abusolutely length (not related to length of arrow).}

\item{angle}{angle from the shaft of the arrow to the edge of the arrow head.}

\item{...}{graphical parameters such as the line characteristics.}
}
\description{
Draws arrows between pairs of points with filling arrow head.
}
