% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_process.R
\name{convert2Gray}
\alias{convert2Gray}
\title{Converts a image to gray scaled.}
\usage{
convert2Gray(img)
}
\arguments{
\item{img}{target image formed \code{array}.}
}
\value{
converted image formed \code{matrix}.
}
\description{
Converts a image to gray scaled.
}
