% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_process.R
\name{noizeSaltPapper}
\alias{noizeSaltPapper}
\title{Apply noize to the image.}
\usage{
noizeSaltPapper(img, salt.rate = 0.01, papper.rate = 0.01)
}
\arguments{
\item{img}{applied \code{matrix} or \code{array} expressing the image.}

\item{salt.rate}{salt (bleaching) rate on the each pixels.}

\item{papper.rate}{papper (darking) rate on the each pixels.}
}
\value{
image applied noize.
}
\description{
Apply noize to the image.
}
