% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{showImage}
\alias{showImage}
\title{Plot a image with `matrix` or `array` formed.}
\usage{
showImage(img, ...)
}
\arguments{
\item{img}{image `matrix` or `array`.}

\item{...}{optional arguments to be passed to [plot()] function.}
}
\description{
Plot a image with `matrix` or `array` formed.
}
\examples{
img <- jpeg::readJPEG(paste(R.home(),"/doc/html/logo.jpg",sep=""))
showImage(img)
}
