% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{showImageList}
\alias{showImageList}
\title{Plot many listed image at once.}
\usage{
showImageList(lst, row_col = 3, plot.mar = c(2, 2, 2, 1), zero.mar = F,
  ...)
}
\arguments{
\item{lst}{image list.}

\item{row_col}{number of images that is plot at onece.
if \code{row_col == 3}, 9 images plot as 3 times 3 layout.
if \code{row_col == c(4,5)}, 20 images plot as 4 times 5 layout.}

\item{plot.mar}{margins of each images.
default is \code{c(2, 2, 2, 1), indicates top, left, bottom, right, respectively.}}

\item{zero.mar}{if \code{T}, plot.mar is set as \code{c(0, 0, 0, 0)}.}

\item{...}{optional argments for plotting image.}
}
\description{
Plot many listed image at once.
}
