% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{showRawImage}
\alias{showRawImage}
\title{Wlot image without axis.}
\usage{
showRawImage(img, ...)
}
\arguments{
\item{img}{image.}

\item{...}{optional argments to be passed to [plot()] function.}
}
\description{
Wlot image without axis.
}
\seealso{
[showImage()]
}
