% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{showImage}
\alias{showImage}
\title{Plot a \code{matrix} or \code{array} expressing a image.}
\usage{
showImage(img, ...)
}
\arguments{
\item{img}{showed \code{matrix} or \code{array} expressing a image.}

\item{...}{arguments to be passed to plot function.}
}
\description{
Plot a \code{matrix} or \code{array} expressing a image.
}
\examples{
img <- jpeg::readJPEG(paste(R.home(),"/doc/html/logo.jpg",sep=""))
showImage(img)

}
